fix index for products created after the deleted product

+++ src/app/api/deleteProduct/route.js
@@ -7,6 +7,18 @@ export async function DELETE(req) {
   const url = new URL(req.url);
 
   await connectDB();
+  const productToDelete = await Product.findById({
+    _id: url.searchParams.get("id"),
+  });
+  if (!productToDelete) {
+    return NextResponse.json({ error: "product for deletion not found" });
+  }
+  const date = new Date(productToDelete.createdAt);
+  const products = await Product.find({ createdAt: { $gt: date } });
+  products.forEach(async (product) => {
+    product.index--;
+    await product.save();
+  });
   const result = await Product.deleteOne({ _id: url.searchParams.get("id") });
 
   if (result.deletedCount !== 1) {
